Implementacion y actualización del dominio Facturación
Se actualizaron los servicios y controladores de este dominio, además de implementar logs y exceptions.
